- Cleaned up analytics calls, made them all use a single “General” category. - Removed redundant call for when the colour button was pressed. - Add a call for when the user taps the ratings button.

+++ Toastmasters Timer/libraries/SGDeepCopy/SGDeepCopy.m
@@ -13,7 +13,7 @@
 @implementation NSArray (SGDeepCopy)
 
 - (NSArray*) deepCopy {
-    unsigned int count = [self count];
+    NSInteger count = [self count];
     id cArray[count];
     
     for (unsigned int i = 0; i < count; ++i) {
@@ -29,7 +29,7 @@
     return ret;
 }
 - (NSMutableArray*) mutableDeepCopy {
-    unsigned int count = [self count];
+    NSInteger count = [self count];
     id cArray[count];
     
     for (unsigned int i = 0; i < count; ++i) {
@@ -67,7 +67,7 @@
 @implementation NSDictionary (SGDeepCopy)
 
 - (NSDictionary*) deepCopy {
-    unsigned int count = [self count];
+    NSInteger count = [self count];
     id cObjects[count];
     id cKeys[count];
     
@@ -96,7 +96,7 @@
 }
 
 - (NSMutableDictionary*) mutableDeepCopy {
-    unsigned int count = [self count];
+    NSInteger count = [self count];
     id cObjects[count];
     id cKeys[count];
